fix bug

+++ src/main/scala/com/cpc/spark/qukan/interest/CheckStudentTag.scala
@@ -97,7 +97,7 @@ object CheckStudentTag {
       """
         |select distinct member_id, info from  gobblin.qukan_member_zfb_log
         | where update_time >= "2018-01-01" and member_id = %s
-      """.stripMargin.format(zfb.first().get(0).toLong)
+      """.stripMargin.format(zfb.first().get(0))
